+++ src/lib.rs
@@ -39,7 +39,7 @@ use self::backends::ConnectionPool;
 pub use self::batch_info::BatchInfo;
 pub use self::category::Category;
 pub use self::cluster::*;
-pub use self::collections::{Indexable, Indexed};
+pub use self::collections::Indexable;
 pub(crate) use self::collections::{IndexedMap, IndexedMapUpdate, IterableMap, Map};
 pub use self::column_statistics::*;
 pub use self::event::EventKind;