+++ src/tables/network.rs
@@ -9,8 +9,8 @@ use serde::{Deserialize, Serialize};
 
 use super::UniqueKey;
 use crate::{
-    HostNetworkGroup, Indexable, Indexed, IndexedMap, IndexedMapUpdate, IndexedTable, Iterable,
-    types::FromKeyValue,
+    HostNetworkGroup, Indexable, IndexedMap, IndexedMapUpdate, IndexedTable, Iterable,
+    collections::Indexed, types::FromKeyValue,
 };
 
 #[derive(Clone, PartialEq, Debug)]